+++ src/finaletoolkit/frag/frag_length.py
@@ -5,15 +5,124 @@ from sys import stdout, stderr
 from shutil import get_terminal_size
 from multiprocessing import Pool
 import gzip
+from functools import partial
 
 import numpy as np
 import pysam
 import tqdm
+import matplotlib.pyplot as plt
+from matplotlib.ticker import FuncFormatter
 
 from finaletoolkit.utils.utils import (
     _get_intervals, frag_generator
 )
 from finaletoolkit.utils.cli_hist import _cli_hist
+
+
+def plot_histogram(
+        data_dict,
+        num_bins,
+        histogram_path="./frag_length_bins_histogram.png",
+        stats=None):
+    """
+    Generates a histogram from `frag_length_bins` results.
+    """
+    keys = list(data_dict.keys())
+    values = list(data_dict.values())
+
+    fig_size = (6, 4)
+    font_size = 12
+    plt.figure(figsize=fig_size, dpi=1000)
+    plt.hist(keys, bins=num_bins, weights=values, color='salmon', edgecolor='white', linewidth=0.1)
+    plt.xlabel("Fragment Size (bp)", fontsize=font_size*0.8)
+    plt.ylabel("Number of Fragments", fontsize=font_size*0.8)
+    plt.xticks(fontsize=font_size * 0.7)
+    plt.yticks(fontsize=font_size * 0.7)
+
+    def format_ticks(value, pos):
+        if value >= 1e6:
+            return '{:1.0f}M'.format(value * 1e-6)
+        elif value >= 1e3:
+            return '{:1.0f}K'.format(value * 1e-3)
+        else:
+            return '{:1.0f}'.format(value)
+
+    plt.gca().yaxis.set_major_formatter(FuncFormatter(format_ticks))
+    plt.gca().spines['top'].set_visible(False)
+    plt.gca().spines['right'].set_visible(False)
+
+    if stats:
+        stats_str = "\n".join([f"{stat[0]}: {stat[1]}" for stat in stats])
+        plt.text(0.95, 0.95, stats_str, transform=plt.gca().transAxes,
+                 fontsize=font_size * 0.6, verticalalignment='top', horizontalalignment='right',
+                 bbox=dict(facecolor='white', alpha=0.7, edgecolor='none'))
+
+    plt.tight_layout()
+    plt.savefig(histogram_path)
+
+
+def _distribution_from_gen(generator):
+    value_counts = {}
+    for fragment in generator:
+        length_of_fragment = fragment[2] - fragment[1]
+        if length_of_fragment in value_counts:
+            value_counts[length_of_fragment] += 1
+        else:
+            value_counts[length_of_fragment] = 1
+    return value_counts
+
+
+def _find_median(val_freq_dict):
+    val = np.array(list(val_freq_dict.keys()))
+    freq = np.array(list(val_freq_dict.values()))
+    ord = np.argsort(val)
+    val = val[ord]
+    freq = freq[ord]
+    cdf = np.cumsum(freq)
+    
+    total_count = cdf[-1]
+    if total_count % 2 == 1:
+        median_index = np.searchsorted(cdf, total_count // 2)
+        median_val = val[median_index]
+        return float(median_val)
+    else:
+        median_indices = np.searchsorted(cdf, [total_count // 2, total_count // 2 + 1])
+        median_vals = val[median_indices]
+        median_val = np.mean(median_vals)
+        return float(median_val)
+
+
+def _frag_length_stats(
+    input_file: Union[str, pysam.AlignmentFile],
+    contig: str,
+    start: int,
+    stop: int,
+    name: str,
+    min_length: int,
+    max_length: int,
+    intersect_policy: str,
+    quality_threshold: int,
+    verbose: Union[bool, int]
+):
+    frag_gen = frag_generator(input_file, contig, quality_threshold, start, stop, min_length, max_length, intersect_policy, verbose)
+    frag_len_dict=_distribution_from_gen(frag_gen)
+
+    if sum(frag_len_dict.values())==0:
+        mean, median, stdev, minimum, maximum = 5*[-1]
+    else:
+        mean = sum(value * count for value, count in frag_len_dict.items()) / sum(frag_len_dict.values())
+        median = _find_median(frag_len_dict)
+        variance = sum(count * ((value - mean) ** 2) for value, count in frag_len_dict.items()) / sum(frag_len_dict.values())
+        stdev = variance ** 0.5
+        minimum = min(frag_len_dict.keys())
+        maximum = max(frag_len_dict.keys())
+
+    return contig, start, stop, name, mean, median, stdev, minimum, maximum
+
+
+def _frag_length_stats_star(partial_frag_stat, interval):
+    contig, start, stop, name = interval
+    return partial_frag_stat(contig=contig, start=start, stop=stop, name=name)
 
 
 def frag_length(
@@ -113,117 +222,20 @@ def frag_length(
 
     return lengths
 
-def _cli_frag_length(
-        input_file: Union[str, pysam.AlignmentFile, pysam.TabixFile],
-        contig: str=None,
-        start: int=None,
-        stop: int=None,
-        intersect_policy: str="midpoint",
-        output_file: str=None,
-        quality_threshold: int=30,
-        verbose: bool=False
-    ) -> np.ndarray:
-    """
-    frag_length optimized for pipelining.
 
-    Parameters
-    ----------
-    input_file : str or pysam.AlignmentFile
-        BAM, SAM, or CRAM file containing paired-end fragment reads or
-        its path. `AlignmentFile` must be opened in read mode.
-    contig : string, optional
-        Contig or chromosome to get fragments from
-    start : int, optional
-        0-based left-most coordinate of interval
-    stop : int, optional
-        1-based right-most coordinate of interval
-    intersect_policy : str, optional
-        Specifies what policy is used to include fragments in the
-        given interval. Default is "midpoint". Policies include:
-        - midpoint: the average of end coordinates of a fragment lies
-        in the interval.
-        - any: any part of the fragment is in the interval.
-    output_file : string, optional
-    quality_threshold : int, optional
-    verbose : bool, optional
-
-    Returns
-    -------
-    lengths : numpy.ndarray
-        `ndarray` of fragment lengths from file and contig if
-        specified.
-    """
-    if (verbose):
-        start_time = time.time()
-        stderr.write("Finding frag lengths.\n")
-
-    lengths = []    # list of fragment lengths
-    
-    frag_gen = frag_generator(
-        input_file=input_file,
-        contig=contig,
-        quality_threshold=quality_threshold,
-        start=start,
-        stop=stop,
-        fraction_low=0,
-        fraction_high=1000000000,   #TODO: allow to have None
-        intersect_policy=intersect_policy,
-        verbose=verbose,
-    )
-
-    # check if output specified
-    if (type(output_file) == str):
-        if output_file.endswith(".bin"): # binary file
-
-            for contig, frag_start, frag_stop, _, _ in frag_gen:
-                lengths.append(frag_stop - frag_start)
-
-            # convert to array
-            lengths = np.array(lengths, dtype=np.int32)
-
-            # write to file
-            with open(output_file, 'wt') as out:
-                lengths.tofile(out)
-
-        elif output_file == '-':
-            for contig, frag_start, frag_stop, _, _ in frag_gen:
-                stdout.write(f'{frag_stop - frag_start}\n')
-
-        else:   # all other file types
-            with open(output_file, 'wt') as out:
-                for contig, frag_start, frag_stop, _, _ in frag_gen:
-                    out.write(f'{frag_stop - frag_start}\n')
-            
-
-    elif (output_file is not None):
-        raise TypeError(
-            f'output_file is unsupported type "{type(input_file)}". '
-            'output_file should be a string specifying the path of the file '
-            'to write output scores to.'
-            )
-
-    if (verbose):
-        end_time = time.time()
-        stderr.write(
-            f'frag_length took {end_time - start_time} s to complete\n'
-        )
-
-    return None
-
-
-# NOTE: I'm not sure what contig by contig was supposed to be.
-# It doesn't do anything.
 def frag_length_bins(
     input_file: Union[str, pysam.AlignmentFile],
     contig: str=None,
     start: int=None,
     stop: int=None,
-    bin_size: int=None,
+    min_length: int=0,
+    max_length: int=None,
+    bin_size: int=1,
     output_file: str=None,
-    contig_by_contig: bool=False,
     histogram: bool=False,
     intersect_policy: str="midpoint",
     quality_threshold: int=30,
+    histogram_path: str=None,
     verbose: Union[bool, int]=False
 ) -> tuple[np.ndarray, np.ndarray]:
     """
@@ -239,7 +251,6 @@ def frag_length_bins(
     stop : int, optional
     bin_size : int, optional
     output_file : str, optional
-    contig_by_contig: bool, optional
     histogram: bool, optional
     intersect_policy : str, optional
         Specifies what policy is used to include fragments in the
@@ -263,7 +274,6 @@ def frag_length_bins(
             stop: {stop}
             bin_size: {bin_size}
             output_file: {output_file}
-            contig_by_contig: {contig_by_contig}
             histogram: {histogram}
             quality_threshold: {quality_threshold}
             verbose: {verbose}
@@ -271,53 +281,51 @@ def frag_length_bins(
         )
         start_time = time.time()
 
-    # generating fragment lengths
-    frag_lengths = frag_length(
-        input_file=input_file,
-        contig=contig,
-        start=start,
-        stop=stop,
-        intersect_policy=intersect_policy,
-        quality_threshold=quality_threshold,
-        verbose=verbose-1 if verbose>1 else 0
+    if verbose:
+        stderr.write("Generating fragment dictionary. \n")
+
+    
+    frag_gen = frag_generator(
+        input_file, contig, quality_threshold, start, stop, min_length,
+        max_length, intersect_policy, verbose)
+    
+    frag_len_dict = _distribution_from_gen(frag_gen)
+
+    mean = (sum(value * count for value, count in frag_len_dict.items())
+            / sum(frag_len_dict.values()))
+    variance = (sum(count * ((value - mean) ** 2)
+                    for value, count in frag_len_dict.items())
+                / sum(frag_len_dict.values())
     )
     # get statistics
     stats = []
-    stats.append(('mean', np.mean(frag_lengths)))
-    stats.append(('median', np.median(frag_lengths)))
-    stats.append(('stdev', np.std(frag_lengths)))
-    stats.append(('min', np.min(frag_lengths)))
-    stats.append(('max', np.max(frag_lengths)))
+    stats.append(('mean', mean))
+    stats.append(('median', _find_median(frag_len_dict)))
+    stats.append(('stdev', variance ** 0.5))
+    stats.append(('min', min(frag_len_dict.keys())))
+    stats.append(('max', max(frag_len_dict.keys())))
 
-    # generating bins and counts
-    if bin_size is None:
-        if histogram:
-            term_width, term_height = get_terminal_size((80, 24))
-            n_bins = (term_width - 24)
-
-            bin_start = np.min(frag_lengths)
-            bin_stop = np.max(frag_lengths)
-
-            bin_size = round((bin_stop - bin_start) / n_bins)
-        else:
-            bin_size = 5
-
-    bin_start = np.min(frag_lengths)
-    bin_stop = np.max(frag_lengths)
+    bin_start = min(frag_len_dict.keys())
+    bin_stop = max(frag_len_dict.keys())
     n_bins = (bin_stop - bin_start) // bin_size
+    bins = np.arange(bin_start, bin_stop+bin_size, bin_size)
+
 
     bins = np.arange(bin_start, bin_stop, bin_size)
     counts = []
 
-    # generate histogram
-    for bin in bins:
-        count = np.sum(
-            (frag_lengths >= bin)
-            * (frag_lengths < (bin + bin_size))
-        )
-        counts.append(count)
-    bins = np.append(bins, stop)
+    # generate histogram data
+    for i in tqdm.tqdm(range(n_bins+1),
+                   disable=not verbose, desc="Binning fragments..."):
+        bin_lower = bin_start + i * bin_size
+        bin_upper = bin_start + (i + 1) * bin_size
+        bin_count = sum(count for length, count in frag_len_dict.items()
+                        if bin_lower <= length < bin_upper)
+        if bin_count == None:
+            bin_count = 0 
+        counts.append(bin_count)
 
+    # write results to output
     if output_file is not None:
         try:
             out_is_file = False
@@ -330,26 +338,22 @@ def frag_length_bins(
                 out_is_file = True
                 out = open(output_file, 'w')
 
-            if histogram:
-                _cli_hist(bins, counts, n_bins, stats, out)
+            out.write('min\tmax\tcount\n')
+            for bin, count in zip(bins, counts):
+                out.write(f'{bin}\t{bin+bin_size}\t{count}\n')
+                
+            if histogram_path!=None:
+                plot_histogram(frag_len_dict, num_bins=n_bins,
+                               histogram_path=histogram_path, stats=stats)
 
-            else:
-                out.write('min\tmax\tcount\n')
-                for bin, count in zip(bins, counts):
-                    out.write(f'{bin}\t{bin+bin_size}\t{count}\n')
         finally:
             if out_is_file:
                 out.close()
-    elif histogram:
-        if contig is not None:
-            if start is None:
-                start = ""
-            if stop is None:
-                stop = ""
-            title = f'Fragment Lengths for {contig}:{start}-{stop}'
-        else:
-            title = 'Fragment Lengths'
-        _cli_hist(bins, counts, n_bins, stats, stdout, title)
+
+    # generate histogram figure
+    elif histogram_path!=None:
+        plot_histogram(frag_len_dict, num_bins=n_bins,
+                       histogram_path=histogram_path, stats=stats)
 
     if verbose:
         stop_time = time.time()
@@ -365,46 +369,48 @@ def _frag_length_stats(
     contig: str,
     start: int,
     stop: int,
+    min_length: int,
+    max_length: int,
     name: str,
     intersect_policy: str,
     quality_threshold: int,
     verbose: Union[bool, int]
 ):
-    # generating fragment lengths
-    frag_lengths = frag_length(
-        input_file=input_file,
-        contig=contig,
-        start=start,
-        stop=stop,
-        intersect_policy=intersect_policy,
-        quality_threshold=quality_threshold,
-        verbose=verbose
-    )
-    if frag_lengths.shape[0] == 0:
-        mean, median, stdev, min, max = 5*[-1]
+    frag_gen = frag_generator(input_file, contig, quality_threshold, start,
+                              stop, min_length, max_length, intersect_policy,
+                              verbose)
+    frag_len_dict=_distribution_from_gen(frag_gen)
+    if sum(frag_len_dict.values())==0:
+        mean, median, stdev, minimum, maximum = 5*[-1]
     else:
-        mean = np.mean(frag_lengths)
-        median = np.median(frag_lengths)
-        stdev = np.std(frag_lengths)
-        min = np.min(frag_lengths)
-        max = np.max(frag_lengths)
+        mean = (sum(value * count for value, count in frag_len_dict.items())
+                / sum(frag_len_dict.values()))
+        median = _find_median(frag_len_dict)
+        variance = sum(count * ((value - mean) ** 2) for value, count
+                       in frag_len_dict.items()) / sum(frag_len_dict.values())
+        stdev = variance ** 0.5
+        minimum = min(frag_len_dict.keys())
+        maximum = max(frag_len_dict.keys())
 
-    return name, contig, start, stop, mean, median, stdev, min, max
+    return contig, start, stop, name, mean, median, stdev, minimum, maximum
 
 
-def _frag_length_stats_star(args):
-    return _frag_length_stats(*args)
+def _frag_length_stats_star(partial_frag_stat, interval):
+    contig, start, stop, name = interval
+    return partial_frag_stat(contig=contig, start=start, stop=stop, name=name)
 
 
 def frag_length_intervals(
     input_file: Union[str, pysam.AlignmentFile],
     interval_file: str,
-    output_file: str=None,
+    output_file: str|None=None,
+    min_length: int|None=0,
+    max_length: int|None=None,
     quality_threshold: int=30,
     intersect_policy: str="midpoint",
     workers: int=1,
     verbose: Union[bool, int]=False
-):
+)->list[tuple[str, int, int, str, float, float, int, int]]:
     """
     Takes fragments from BAM file and calculates fragment length
     statistics for each interval in a BED file. If output specified,
@@ -421,6 +427,7 @@ def frag_length_intervals(
 
     Returns
     -------
+    results: list of (contig, start, stop, name, mean, median, stdev, min, max)'
     """
     if verbose:
         stderr.write(
@@ -428,6 +435,8 @@ def frag_length_intervals(
             input_file: {input_file}
             interval_file: {interval_file}
             output_file: {output_file}
+            min_length: {min_length}
+            max_length: {max_length}
             quality_threshold: {quality_threshold}
             workers: {workers}
             verbose: {verbose}
@@ -435,64 +444,55 @@ def frag_length_intervals(
         )
         start_time = time.time()
 
-    # read interval_file into list of tuples
-    intervals = _get_intervals(
-        input_file,
-        interval_file,
-        intersect_policy=intersect_policy,
-        quality_threshold=quality_threshold,
-        verbose=verbose
-    )
-
-    interval_len = len(intervals)
     if verbose:
-        stderr.write(f'{interval_len} intervals read!\n')
-
+        stderr.write('Creating process pool.\n')
     try:
-        # create pool and submit intervals into processess
-        pool = Pool(workers)
-        iter_results = pool.imap(
-            _frag_length_stats_star,
-            tqdm.tqdm(
-                intervals,
-                desc='Filling process pool',
-                position=0
-            ) if verbose else intervals,
-            round(interval_len/workers/2+1)
-        )
-        # write to output
-        try:
-            output_is_file = False
-            if output_file == '-':
-                out = stdout
-            else:
-                output_is_file = True
-                out = open(output_file, 'w')
-            # header
-            out.write(
-                'name\tcontig\tstart\tstop\tmean\tmedian\tstdev\tmin\tmax\n'
-            )
-            for result in tqdm.tqdm(
-                iter_results,
-                total=interval_len,
-                desc='Writing to out',
-                position=2
-            ) if verbose else iter_results:
-                out.write('\t'.join([str(item) for item in result]))
-                out.write('\n')
-        finally:
-            if output_is_file:
-                out.close()
-
-        results = [result for result in iter_results]
+        pool = Pool(processes=workers)
+        if verbose:
+            stderr.write('Reading intervals.\n')
+        intervals = _get_intervals(interval_file)
+        
+        partial_frag_stat = partial(_frag_length_stats, input_file=input_file, min_length=min_length, max_length=max_length, intersect_policy=intersect_policy, quality_threshold=quality_threshold, verbose=verbose)
+        results = pool.map(partial(_frag_length_stats_star, partial_frag_stat), intervals, chunksize=max(len(intervals)//workers, 1))
+        if verbose:
+            tqdm.write('Retrieving fragment statistics for file\n')
+        output_is_file = False
+        if output_file != None:
+            if verbose:
+                tqdm.write('Writing results to output. \n')
+            try:
+                if (output_file.endswith('.bed')
+                    or output_file.endswith('.bedgraph')
+                ):
+                    output_is_file = True
+                    output = open(output_file, 'w')
+                elif output_file.endswith('.bed.gz'):
+                    output = gzip.open(output_file, 'w')
+                    output_is_file = True
+                elif output_file == '-':
+                    output = stdout
+                else:
+                    raise ValueError(
+                        'The output file should have .bed or .bed.gz as as '
+                        'suffix.'
+                    )
+                output.write('contig\tstart\tstop\tname\tmean\tmedian\t'
+                             'stdev\tmin\tmax\n') # type: ignore
+                output.write(
+                    '\n'.join('\t'.join(str(element) for element in 
+                                        item) for item in results)) # type: ignore
+                output.write('\n') # type: ignore
+            finally:
+                if output_is_file:
+                    output.close()
 
     finally:
         pool.close()
-
+        
     if verbose:
         stop_time = time.time()
         stderr.write(
-            f'frag_length_intervals took {stop_time-start_time} s\n'
+            f'Calculating fragment length statistics for intervals took {stop_time-start_time} s\n'
         )
 
     return results